requirements.txt
pillow==10.4.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.2
vobject==0.9.7
segno==1.6.1
